Created the normalisation function
